requirements.txt
crewai
textwrap3
streamlit
typing
crewai_tools
requests
markdownify
bs4
python-dotenv
markdown2